change(api)!: use new consolidated routes for api spec and docs
To have a consistent and more intuitive way of accessing the OpenAPI
specification and the Swagger UI as a documentation of this API, accross
all base projects we will adopt these new routes on all projects.

This is not strictly a breaking change in terms of the backend, but if
a frontend is already consuming the old routes, one has to adapt the
config accordingly to the new API spec.